Rend2: Reduce vertex data size for g2 models
Bone indices and bone weights are now stored in uint8 x 4 instead of
float x 4.

Vertex data size before this commit was:
position = 12 bytes (3 x f32)
normals = 4 bytes (1 x uint1010102)
texcoords = 8 bytes (2 x f32)
weights = 16 bytes (4 x f32)
bone indices = 16 bytes (4 x f32)
tangents = 4 bytes (1 x uint1010102)
TOTAL = 60 bytes

After this commit:
weights = 4 bytes
indices = 4 bytes
TOTAL = 36 bytes

60 bytes -> 36 bytes = 40% vertex data size reduction

+++ codemp/rd-rend2/glsl/fogpass_vp.glsl
@@ -7,7 +7,7 @@ in vec2 attr_TexCoord0;
 in vec3 attr_Position2;
 in vec3 attr_Normal2;
 #elif defined(USE_SKELETAL_ANIMATION)
-in vec4 attr_BoneIndexes;
+in uvec4 attr_BoneIndexes;
 in vec4 attr_BoneWeights;
 #endif
 
@@ -200,7 +200,7 @@ void main()
 
 	for (int i = 0; i < 4; i++)
 	{
-		int boneIndex = int(attr_BoneIndexes[i]);
+		uint boneIndex = attr_BoneIndexes[i];
 
 		mat4 boneMatrix = mat4(
 			vec4(u_BoneMatrices[boneIndex][0], 0.0),
